# internal/handler/auth_handler.go
package handler

import (
	"booking-app/internal/dto"
	"booking-app/internal/service"
	"net/http"
	"booking-app/internal/response"
	customValidator "booking-app/internal/validator"
	"github.com/gin-gonic/gin"
)

type AuthHandler struct {
	authService service.AuthService
}

func NewAuthHandler(
	authService service.AuthService,
) *AuthHandler {

	return &AuthHandler{
		authService: authService,
	}
}

// Register godoc
//
// @Summary Register user
// @Description Register a new user
// @Tags Authentication
// @Accept json
// @Produce json
// @Param request body dto.RegisterRequest true "Register Request"
// @Success 201 {object} map[string]interface{}
// @Failure 400 {object} map[string]interface{}
// @Router /auth/register [post]
func (h *AuthHandler) Register(
	c *gin.Context,
) {

	var req dto.RegisterRequest
	ctx := c.Request.Context()
	if err := c.ShouldBindJSON(
		&req,
	); err != nil {

		c.JSON(
			http.StatusBadRequest,
			gin.H{
				"error": err.Error(),
			},
		)
		return
	}

	err :=
		customValidator.Validate.
			Struct(req)

	if err != nil {

		response.Error(
			c,
			http.StatusBadRequest,
			customValidator.
				FormatValidationError(
					err,
				),
		)

		return
	}

	err = h.authService.Register(
		ctx,
		req,
	)

	if err != nil {

		c.JSON(
			http.StatusBadRequest,
			gin.H{
				"error": err.Error(),
			},
		)

		return
	}

	response.Success(
		c,
		http.StatusCreated,
		"user registered",
		nil,
	)
}

// Login godoc
//
// @Summary Login user
// @Description Login with email and password
// @Tags Authentication
// @Accept json
// @Produce json
// @Param request body dto.LoginRequest true "Login Request"
// @Success 200 {object} map[string]interface{}
// @Failure 401 {object} map[string]interface{}
// @Router /auth/login [post]
func (h *AuthHandler) Login(
	c *gin.Context,
) {

	var req dto.LoginRequest
	ctx := c.Request.Context()
	if err := c.ShouldBindJSON(
		&req,
	); err != nil {

		c.JSON(
			http.StatusBadRequest,
			gin.H{
				"error": err.Error(),
			},
		)

		return
	}

	err :=
		customValidator.Validate.
			Struct(req)
	if err != nil {

		response.Error(
			c,
			http.StatusBadRequest,
			customValidator.
				FormatValidationError(
					err,
				),
		)

		return
	}

	token, err :=
		h.authService.Login(ctx, req)

	if err != nil {

		c.JSON(
			http.StatusUnauthorized,
			gin.H{
				"error": err.Error(),
			},
		)

		return
	}

	response.Success(
		c,
		http.StatusOK,
		"login success",
		gin.H{
			"access_token": token,
		},
	)
}